Better flame graph generation for dashboard
The existing node.js libraries for generating flame graphs are not well
maintained and have some issues.  The existing library would sometimes
run into stack overflows when producing the flame graph.
https://github.com/brendangregg/FlameGraph has the canonical
implementation, but it is in Perl.  Given the current state of the
node.js libraries the dashboard now shells out to run this perl script.